+++ abipy/panels/gsr.py
@@ -9,6 +9,7 @@ from .core import PanelWithElectronBands, PanelWithEbandsRobot
 
 class GsrFilePanel(PanelWithElectronBands):
     """
+    Panel with widgets to interact with a |GsrFile|.
     """
     def __init__(self, gsr, **params):
         super().__init__(**params)
@@ -25,6 +26,7 @@ class GsrFilePanel(PanelWithElectronBands):
         app(("Summary", pn.Row(bkw.PreText(text=self.gsr.to_string(verbose=self.verbose),
                                sizing_mode="scale_both"))))
         app(("e-Bands", pn.Row(self.get_plot_ebands_widgets(), self.on_plot_ebands_btn)))
+
         # Add DOS tab only if k-sampling.
         if self.gsr.ebands.kpoints.is_ibz:
             app(("e-DOS", pn.Row(self.get_plot_edos_widgets(), self.on_plot_edos_btn)))
@@ -55,6 +57,7 @@ class GsrRobotPanel(PanelWithEbandsRobot):
         app(("Summary", pn.Row(bkw.PreText(text=self.robot.to_string(verbose=self.verbose),
                                sizing_mode="scale_both"))))
         app(("e-Bands", pn.Row(self.get_ebands_plotter_widgets(), self.on_ebands_plotter_btn)))
+
         # Add e-DOS tab only if all ebands have k-sampling.
         if all(abifile.ebands.kpoints.is_ibz for abifile in self.robot.abifiles):
             app(("e-DOS", pn.Row(self.get_edos_plotter_widgets(), self.on_edos_plotter_btn)))