+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Release 0.8.0: xxxx-xx-xx
     * Add examples for quasi-harmonic calculations and post-processing tools
     * Add support for JSON files (including MSONable format) in abiopen.py
     * Improved support for EPH calculations.
+    * Add `primitive` command to `abistruct.py` to get primite structure from spglib
 
 Release 0.7.0: 2019-10-18